requirements.txt
grpcio==1.66.0
grpcio-tools==1.66.0
protobuf==5.27.2
fastapi==0.115.0
uvicorn==0.30.0
orjson==3.10.7
requests==2.32.3
beautifulsoup4==4.12.3
